// apps/frontend/types/relatorios.ts
export interface IRelatorioFiltrosState {
    tipoRelatorio: string | null;
    extensaoArquivo: string | null;
    periodoString: string | null;
    dataInicial: Date | string | null;
    dataFinal: Date | string | null;
    anoInicial: string | null | Date;
    anoFinal: string | null | Date;
}